Move case classes into models folder
