+++ handlers/soft-opt-in-consent-setter/src/main/scala/com/gu/soft_opt_in_consent_setter/models/SFSubscription.scala
@@ -1,4 +1,4 @@
-package com.gu.soft_opt_in_consent_setter
+package com.gu.soft_opt_in_consent_setter.models
 
 object SFSubscription {
